+++ packages/operators/src/fs/index.ts
@@ -1,7 +1,7 @@
 export * from './ls'
 export * from './readFile'
 export * from './exists'
-export * from './bound'
+export * from './bindings'
 export * from './fromFileStream'
 export * from './streamIntoFile'
 export * from './writeFile'
